chore: remove SettingGrouping module
The original intention of the design of the management part of the official settings module of abp is to be solved by the developer, while the original intention of my extension of this module is that the simple configuration of the developer can quickly realize the setting management. The original intention is different, causing the extension code to be very awkward, so I removed it with tears.
Although it took a lot of effort to develop the SettingGrouping module, I still decided to remove this chicken rib.
It is recommended to use the official contributor mode of abp to develop settings management
A branch named "keep-SettingGrouping-module" was created before removal.
Closes #9 , #8 , #1